// src/generated/models/CloneOrderEndpoint_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CloneOrderEndpoint_Output = {
    /**
     * The dataset ID of the cloned order.
     */
    datasetId?: string | null;
    /**
     * The ID of the newly cloned order.
     */
    orderId: string;
};