// src/generated/models/CreateComplexOrderEndpoint_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateComplexOrderEndpoint_Output = {
    /**
     * The dataset ID of the created order.
     */
    datasetId?: string | null;
    /**
     * The ID of the created order.
     */
    orderId: string;
    /**
     * The campaign ID of the created order.
     */
    campaignId?: string | null;
    /**
     * The pipeline ID of the created order.
     */
    pipelineId: string;
};